Move AVX2 wrapper to utils
