+++ src/lib/block/serpent/serpent_avx2/info.txt
@@ -4,9 +4,6 @@ SERPENT_AVX2 -> 20180824
 
 need_isa avx2
 
-<cc>
-gcc
-clang
-msvc
-icc
-</cc>
+<requires>
+simd_avx2
+</requires>
